fix: Move existing `config` to `config_legacy`
