+++ src/ai/backend/manager/api/wsproxy.py
@@ -15,7 +15,7 @@ from aiohttp import WSCloseCode, web
 
 from ai.backend.logging import BraceStyleAdapter
 
-from ..config import DEFAULT_CHUNK_SIZE
+from ..config_legacy import DEFAULT_CHUNK_SIZE
 
 log = BraceStyleAdapter(logging.getLogger(__spec__.name))
