+++ src/ai/backend/manager/idle.py
@@ -84,7 +84,7 @@ if TYPE_CHECKING:
 
     from ai.backend.common.types import AgentId, KernelId, SessionId
 
-    from .config import SharedConfig
+    from .config_legacy import SharedConfig
     from .models.utils import ExtendedAsyncSAEngine as SAEngine
 
 log = BraceStyleAdapter(logging.getLogger(__spec__.name))
