+++ src/ai/backend/manager/services/metric/container_metric.py
@@ -15,7 +15,7 @@ from ai.backend.common.metrics.types import (
 )
 from ai.backend.common.types import HostPortPair
 from ai.backend.logging import BraceStyleAdapter
-from ai.backend.manager.config import SharedConfig
+from ai.backend.manager.config_legacy import SharedConfig
 
 from .actions.container import (
     ContainerMetricAction,
