+++ src/ai/backend/manager/services/vfolder/services/invite.py
@@ -10,7 +10,7 @@ from sqlalchemy.orm import contains_eager
 from ai.backend.common.types import (
     VFolderHostPermission,
 )
-from ai.backend.manager.config import SharedConfig
+from ai.backend.manager.config_legacy import SharedConfig
 from ai.backend.manager.models.user import UserRole, UserRow
 from ai.backend.manager.models.utils import ExtendedAsyncSAEngine
 from ai.backend.manager.models.vfolder import (
