+++ tests/manager/api/test_config.py
@@ -6,7 +6,7 @@ import pytest
 @pytest.mark.asyncio
 async def test_register_myself(shared_config, mocker):
     instance_id = "i-test-manager"
-    from ai.backend.manager import config as config_mod
+    from ai.backend.manager import config_legacy as config_mod
 
     mocked_get_instance_id = AsyncMock(return_value=instance_id)
     mocker.patch.object(config_mod, "get_instance_id", mocked_get_instance_id)
